src/backend/semantic/requirements.txt
pymilvus==2.4.0
nats-py==2.7.2
protobuf==4.21.6
grpcio-tools==1.60.0
python-dotenv==1.0.1
webdriver-manager==4.0.1
selenium==4.21.0
beautifulsoup4==4.12.3
langchain-text-splitters==0.2.0
PyMuPDF==1.24.4
pymupdf4llm==0.0.5
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
sqlalchemy-cockroachdb==2.0.2
youtube-transcript-api==0.6.2
pymupdf4llm==0.0.5
PyMuPDFb==1.24.3
mistune==3.0.2
minio==7.2.7



